// src/types/social-card.ts
export interface SocialCardData {
  name: string;
  title?: string;
  website?: string;
  email?: string;
  linkedin?: string;
  github?: string;
  twitter?: string;
  portfolio?: string;
  photoUrl: string;
  phone?: string;
  about?: string;
  interests?: string;
  techSkills?: string;
  publishedWorks?: string;
  latestWorkLinks?: string;
  gradient?: 'blue' | 'purple' | 'teal' | 'orange' | 'green' | 'dark' | 'light';
}